# openapi-generator-spec/accounts/api-spec.yaml
openapi: 3.0.3
info:
  title: Accounts microservice - OpenAPI 3.0
  description: Accounts microservice API Spec
  version: 1.0.0

servers:
  - url: api/v1

paths:
  /accounts:
    get:
      tags:
        - Accounts
      summary: Get all accounts
      description: Get all accounts information
      operationId: getAccounts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiAccount'
        '400':
          description: Invalid request
    post:
      tags:
        - Accounts
      summary: Create an account for current logged in user
      description: Create an account for current logged in user
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiAccount'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiAccount'
        '400':
          description: Invalid request

  /accounts/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Get details of specified account number
      description: Get details of specified account number
      operationId: getAccountDetails
      parameters:
        - in: path
          name: accountNumber
          required: true
          description: account number for which details have to be fetched
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiAccount'
        '400':
          description: Invalid request

  /accounts//balances/{balanceEntry}/{accountNumber}:
    put:
      tags:
        - Balances
      summary: Updates account balance
      description: Credit/Debit from the specified account number
      operationId: updateBalance
      parameters:
        - in: path
          name: balanceEntry
          required: true
          description: Type of entry debit/credit
          schema:
            $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiBalanceEntry'
        - in: path
          name: accountNumber
          required: true
          description: Account number from which the balance has to be debited/credited
          schema:
            type: string
      requestBody:
        description: Balance amount that is to be debited/credited
        required: true
        content:
          application/json:
            schema:
              $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiBalance'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '../shared-components/shared-spec.yaml#/components/schemas/ApiBalance'
        '400':
          description: Invalid request